requirements.txt
aerich==0.6.3
aiofiles==23.1.0
aiomysql==0.1.1
aioredis==2.0.1
aiosqlite==0.17.0
anyio==3.6.1
async-timeout==4.0.2
autopep8==1.7.0
click==8.1.3
dictdiffer==0.9.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.79.0
fastapi-pagination==0.9.3
gunicorn==20.1.0
h11==0.13.0
idna==3.3
iso8601==1.0.2
loguru==0.6.0
passlib==1.7.4
pyasn1==0.4.8
pycodestyle==2.9.1
pydantic==1.9.2
PyMySQL==1.0.2
pypika-tortoise==0.1.6
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.2.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
supervisor==4.2.4
toml==0.10.2
tomlkit==0.11.3
tortoise-orm==0.19.2
typing_extensions==4.3.0
uvicorn==0.18.2
